Update damage_pyro_volcanyro.sp
Upping Gas Passer recharge slightly (24 -> 15 seconds) because throwing it and suiciding recharges it faster than landing a successful explosion against a group.
Also adjusted one of the values on the Pyromancer's Mask as it is not visible in game due to being too small. Unsure if the number adjusted is correct.